score

+++ DFT.h
@@ -6,6 +6,6 @@ extern short TabCos[64];
 extern short TabSin[64];
 extern short LeSignal[64];
 
-long long DFT(short * Signal, int k);
+int DFT(short * Signal, int k);
 
 #endif
